requirements.txt
openai>=1.0.0
requests
Pillow
django
djangorestframework
python-decouple
django-cors-headers
django-filter
djangorestframework-simplejwt
channels
daphne
whitenoise
celery
redis
django-celery-beat
django-celery-results
geopy
reportlab
razorpay
numpy
pandas
scikit-learn
openpyxl
gunicorn
psycopg2-binary